Arcane/requirements.txt
pymongo
pip==23.2.1
bertopic==0.16.0
boto3==1.34.0
botocore==1.34.0
chardet>=3.0.2,<5;
Flask==2.2.2
huggingface-hub==0.16.4
hydra-core==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
psycopg2-binary
psutil
torch==1.10.0
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
smart-open==6.4.0
sagemaker==2.205.0
sshtunnel
tokenizers==0.14
tqdm==4.65.0
transformers==4.35.2
umap-learn==0.5.4
werkzeug==2.2.2
openai
tiktoken
datasets